# contracts/openapi/sysmind-local-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SysMind AI Local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/actions/candidates": {
      "get": {
        "tags": [
          "actions"
        ],
        "summary": "Candidates",
        "operationId": "candidates_api_v1_actions_candidates_get",
        "parameters": [
          {
            "name": "diagnosis_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "title": "Diagnosis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/process-candidates": {
      "get": {
        "tags": [
          "actions"
        ],
        "summary": "Process Candidates",
        "operationId": "process_candidates_api_v1_actions_process_candidates_get",
        "parameters": [
          {
            "name": "diagnosis_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 36,
              "maxLength": 36,
              "title": "Diagnosis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessCandidateListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions": {
      "get": {
        "tags": [
          "actions"
        ],
        "summary": "Recent Actions",
        "operationId": "recent_actions_api_v1_actions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Create Action",
        "operationId": "create_action_api_v1_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/process-close": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Create Process Close",
        "operationId": "create_process_close_api_v1_actions_process_close_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProcessActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{close_action_id}/termination": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Create Process Termination",
        "operationId": "create_process_termination_api_v1_actions__close_action_id__termination_post",
        "parameters": [
          {
            "name": "close_action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Close Action Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{action_id}": {
      "get": {
        "tags": [
          "actions"
        ],
        "summary": "Get Action",
        "operationId": "get_action_api_v1_actions__action_id__get",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{action_id}/confirm": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Confirm Action",
        "operationId": "confirm_action_api_v1_actions__action_id__confirm_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{action_id}/reject": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Reject Action",
        "operationId": "reject_action_api_v1_actions__action_id__reject_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{action_id}/execute": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Execute Action",
        "operationId": "execute_action_api_v1_actions__action_id__execute_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actions/{action_id}/recovery": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Create Recovery",
        "operationId": "create_recovery_api_v1_actions__action_id__recovery_post",
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ActionListResponse"
      },
      "ActionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "diagnosis_id": {
            "type": "string",
            "title": "Diagnosis Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "tool_version": {
            "type": "string",
            "title": "Tool Version"
          },
          "target_name": {
            "type": "string",
            "title": "Target Name"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "recovery_available": {
            "type": "boolean",
            "title": "Recovery Available"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "diagnosis_id",
          "tool_name",
          "tool_version",
          "target_name",
          "source_kind",
          "status",
          "recovery_available",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "ActionResponse"
      },
...
      "CandidateListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CandidateListResponse"
      },
      "CandidateResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "command_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Name"
          },
          "observed_revision": {
            "type": "string",
            "title": "Observed Revision"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "name",
          "source_kind",
          "command_name",
          "observed_revision"
        ],
        "title": "CandidateResponse"
      },
...
      "ConsentResponse": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionResponse"
          },
          "ticket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "action",
          "ticket",
          "expires_at"
        ],
        "title": "ConsentResponse"
      },
...
      "CreateActionRequest": {
        "properties": {
          "diagnosis_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Diagnosis Id"
          },
          "item_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Item Id"
          },
          "observed_revision": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Observed Revision"
          }
        },
        "type": "object",
        "required": [
          "diagnosis_id",
          "item_id",
          "observed_revision"
        ],
        "title": "CreateActionRequest"
      },
      "CreateProcessActionRequest": {
        "properties": {
          "diagnosis_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 36,
            "title": "Diagnosis Id"
          },
          "item_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Item Id"
          },
          "observed_revision": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Observed Revision"
          }
        },
        "type": "object",
        "required": [
          "diagnosis_id",
          "item_id",
          "observed_revision"
        ],
        "title": "CreateProcessActionRequest"
      },
...
      "ExecuteActionRequest": {
        "properties": {
          "ticket": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 20,
            "title": "Ticket"
          }
        },
        "type": "object",
        "required": [
          "ticket"
        ],
        "title": "ExecuteActionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessCandidateListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProcessCandidateResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ProcessCandidateListResponse"
      },
      "ProcessCandidateResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "command_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Name"
          },
          "observed_revision": {
            "type": "string",
            "title": "Observed Revision"
          },
          "cpu_percent": {
            "type": "number",
            "title": "Cpu Percent"
          },
          "memory_percent": {
            "type": "number",
            "title": "Memory Percent"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "name",
          "source_kind",
          "command_name",
          "observed_revision",
          "cpu_percent",
          "memory_percent"
        ],
        "title": "ProcessCandidateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}